% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIMstatus.R
\name{ZIM_status}
\alias{ZIM_status}
\title{Classification of olive tree water status from ZIM probes data}
\usage{
ZIM_status(model, new_data)
}
\arguments{
\item{model}{Indicate the model chosen for the classification of the Pp daily curves.}

\item{new_data}{Indicate the columns of a data frame with the variables obtained with the function "getpoints".}
}
\value{
Tree water status (1, 2 or 3).
}
\description{
Classification of olive tree water status from ZIM probes data
}
